# .mem/openapi.yaml
openapi: 3.0.3
info:
  title: Co-Work API
  description: API documentation for the Co-Work collaborative project management system
  version: 1.0.0
  contact:
    name: Co-Work Team
servers:
  - url: /api
    description: API base path
...
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
      required:
        - error
    Project:
      type: object
      properties:
        id:
          type: integer
          description: Project ID
        name:
          type: string
          description: Project name
        status:
          type: string
          enum: [saved, unsaved]
          description: Project status
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
      required:
        - id
        - name
        - status
        - createdAt
        - updatedAt
    ProjectMember:
      type: object
      properties:
        id:
          type: integer
          description: Member ID
        projectId:
          type: integer
          description: Project ID
        userId:
          type: integer
          description: User ID
        role:
          type: string
          enum: [owner, editor, viewer]
          description: Member role
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
      required:
        - id
        - projectId
        - userId
        - role
        - createdAt
...
paths:
  /projects:
    get:
      tags:
        - Projects
      summary: List projects
      description: Retrieve a list of projects with optional search and pagination
      parameters:
        - name: id
          in: query
          description: Project ID for single project retrieval
          schema:
            type: integer
        - name: limit
          in: query
          description: Maximum number of results (max 100)
          schema:
            type: integer
            default: 10
            maximum: 100
        - name: offset
          in: query
          description: Number of items to skip
          schema:
            type: integer
            default: 0
        - name: search
          in: query
          description: Search term for project name
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Project'
                  - type: array
                    items:
                      $ref: '#/components/schemas/Project'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Projects
      summary: Create a project
      description: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  description: Project name
                status:
                  type: string
                  enum: [saved, unsaved]
                  description: Project status
      responses:
        '201':
          description: Project created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Projects
      summary: Update a project
      description: Update an existing project
      parameters:
        - name: id
          in: query
          required: true
          description: Project ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Project name
                status:
                  type: string
                  enum: [saved, unsaved]
                  description: Project status
      responses:
        '200':
          description: Project updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Projects
      summary: Delete a project
      description: Delete an existing project
      parameters:
        - name: id
          in: query
          required: true
          description: Project ID
          schema:
            type: integer
      responses:
        '200':
          description: Project deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  project:
                    $ref: '#/components/schemas/Project'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /project-members:
    get:
      tags:
        - Project Members
      summary: List project members
      description: Retrieve project members by ID or project
      parameters:
        - name: id
          in: query
          description: Member ID for single member retrieval
          schema:
            type: integer
        - name: projectId
          in: query
          description: Project ID (required for listing members)
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ProjectMember'
                  - type: array
                    items:
                      $ref: '#/components/schemas/ProjectMember'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Member not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Project Members
      summary: Add project member
      description: Add a new member to a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - projectId
                - userId
              properties:
                projectId:
                  type: integer
                  description: Project ID
                userId:
                  type: integer
                  description: User ID
                role:
                  type: string
                  enum: [owner, editor, viewer]
                  default: viewer
                  description: Member role
      responses:
        '201':
          description: Member added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectMember'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Project Members
      summary: Remove project member
      description: Remove a member from a project
      parameters:
        - name: id
          in: query
          required: true
          description: Member ID
          schema:
            type: integer
      responses:
        '200':
          description: Member removed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  member:
                    $ref: '#/components/schemas/ProjectMember'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Member not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...